inserido o card, mas atualizando infinito
