--- Pokedex/src/Components/Header/PokemonList/PokemonList.js
@@ -1,16 +1,0 @@
-import React from "react";
-import Styled from './styled'
-
-const PokemonList = (props) =>{
-    return(
-        <MainContainer>
-            
-
-
-
-        </MainContainer>
-    )
-}
-
-
-export default PokemonList
