+++ Pokedex/src/Global/GlobalState.js
@@ -6,7 +6,23 @@ import GlobalStateContext from '../Contexts/GlobalStateContext'
 const GlobalState = (props) =>{
 
     const [pokemons, setPokemons] = useState([])
-    const [imgPokemon, setImgPokemon]= useState('')
+    const [pokemonDetail, setPokemonDetail] = useState("")
+   
+    const getPokemonDetail = (pokeName) =>{
+
+        axios
+        .get(`${BASE_URL}${pokeName}/`)
+        .then((res)=>{
+            console.log(res)
+            setPokemonDetail(res.data.sprites.front_default)   
+            // console.log(res.data.sprites)
+                     
+        })
+        .catch((err)=>{
+            console.log(err)
+        })
+        console.log(pokemonDetail)
+    }
     
 
     const getPokemons = () =>{
@@ -22,11 +38,13 @@ const GlobalState = (props) =>{
         })
     }
 
+    
+
   
     
-    const states = {pokemons, imgPokemon}
-    const setters ={setPokemons, setImgPokemon}
-    const requests = { getPokemons };
+    const states = {pokemons, pokemonDetail}
+    const setters ={setPokemons, setPokemonDetail}
+    const requests = {getPokemons, getPokemonDetail };
 
     const data = {states, setters, requests };
