+++ Pokedex/src/Pages/Home/Home.js
@@ -1,43 +1,16 @@
 import React, { useState, useEffect, useContext } from 'react'
-import GlobalStateContext from '../../Contexts/GlobalStateContext'
-import { useHistory } from 'react-router-dom'
-import axios from 'axios'
-import { BASE_URL} from '../../Constants/url'
+import PokemonList from '../../Components/PokemonList/PokemonList'
 
 
 const Home = () =>{
-  const { states, setters, requests } = useContext(GlobalStateContext)
-  const history = useHistory()
-  const [pokemon, setPokemon] = useState('')
 
 
-  useEffect(()=>{
-    requests.getPokemons()
-    
-  }, [])
 
-  
-
-const pokemonList = 
-states.pokemons && 
-states.pokemons.map((pokemon)=>{
-  
-    return(
-      <div>
-        <p>{pokemon.name}</p>
-        <div>
-          
-        </div>
-               
-      </div>
-      
-    )
-  })
 
 
   return (
     <div>
-      {pokemonList}
+      <PokemonList/>
    </div>
    
  );
